implemented fixed format writer

+++ crates/spec_parser/src/spec_xml_hard.rs
@@ -185,7 +185,7 @@ pub struct Compound {
 #[xml(tag = "Fixed")]
 pub struct Fixed {
     #[xml(attr = "length")]
-    pub length: u64,
+    pub length: u32,
     #[xml(child = "Bits")]
     pub bits: Vec<Bits>,
 }
@@ -194,11 +194,11 @@ pub struct Fixed {
 #[xml(tag = "Bits")]
 pub struct Bits {
     #[xml(attr = "bit")]
-    pub bit: Option<u64>,
+    pub bit: Option<u32>,
     #[xml(attr = "from")]
-    pub from: Option<u64>,
+    pub from: Option<u32>,
     #[xml(attr = "to")]
-    pub to: Option<u64>,
+    pub to: Option<u32>,
     #[xml(attr = "fx")]
     pub fx: Option<u64>,
     #[xml(attr = "rep")]
@@ -219,8 +219,9 @@ pub struct Bits {
     pub presence: Option<String>,
 }
 
-#[derive(Debug)]
+#[derive(Debug, Default, Copy, Clone, PartialEq, Eq)]
 pub enum Encode {
+    #[default]
     Unsigned,
     Signed,
     Octal,
